get_nc_online

Download and connect climate data from NetCDF file (.nc) retrieved from https://surfobs.climate.copernicus.eu/dataaccess/access_eobs.php

get_nc_online(
  first_year = first_year,
  last_year = last_year,
  sml_chunk = sml_chunk,
  clim_variable = clim_variable,
  statistic = statistic,
  grid_size = grid_size,
  ecad_v = ecad_v
)

Arguments

first_year

a numeric value defining the first year of the time period to extract, 1950 if NULL, default=NULL

last_year

a numeric value defining the last year of the time period to extract, 2014 if NULL, default=NULL

sml_chunk

a character string for specific time period to be downloaded. Chunk available are "2011-2025", "1995-2010", "1980-1994", "1965-1979", and "1950-1964"

clim_variable

a character string defining the daily climate variable to retrieve and extract; "mean temp", "max temp", "min temp", "precipitation", "sea level pressure", "relative humidity", "global radiation", default="mean temp"

statistic

a character string defining the metric to retrieve, "mean" or "spread", where the mean is computed across 100 members of the ensemble and is provided as the "best-guess" fields. The spread is calculated as the difference between the 5th and 95th percentiles over the ensemble to provide a measure indicative of the 90% uncertainty range (see details)

grid_size

numeric, measured in degree defining the resolution of the grid, 0.25 (ca. 27 kilometres in latitude) or 0.1 (ca. 11 kilometres in latitude), default=0.25

ecad_v

ECA&D data version, default = package version.

Value

A connection via the ncdf4 package to a ".nc" file downloaded to the local disk

Details

This function access online ECAD data portal to download the climate ".nc" file to local disk. This is an internal function used within the 'extract_nc_value' function, but can be used independently if needed. If first_year and last_year are not provided, the function extract the full data set. Smaller chunks of about 15 years of the most recent version of the E-OBS dataset can be specified for download can be specified directly with the argument "sm_chunk" (period available are 2011-2025, 1995-2010, 1980-1994, 1965-1979, 1950-1964). For more details about the mean and the spread metric see Cornes et al. (2018) and the guidance on how to use ensemble datasets available from http://surfobs.climate.copernicus.eu/userguidance/use_ensembles.php

Author

Reto Schmucki