write_to_brick

Write the output in a multilayer raster

write_to_brick(x, out = out, ...)

Arguments

x
object obtained from the extrat_nc_value function corresponding to the specific period of interest

out
character string defining the filename to save (default extension is GeoTiff, ".tiff")

...
additional arguments for for writing files, see writeRaster

Details

By default, this function overwrites file with the same name if existing. Layers' names are Date starting with X (e.g. "X2010.01.27")