extract_nc_value

Function to extract climate data from NETCDF file produced by the European Climate Assessment & Dataset for a specific time-period and available at https://surfobs.climate.copernicus.eu/dataaccess/access_eobs.php

extract_nc_value(
  first_year = NULL,
  last_year = NULL,
  local_file = TRUE,
  file_path = NULL,
  sml_chunk = NULL,
  clim_variable = "mean temp",
  statistic = "mean",
  grid_size = 0.25
)

Arguments

first_year
a numeric value defining the first year of the time-period to extract, 1950 if NULL, default=NULL

last_year
a numeric value defining the last year of the time-period to extract, 2014 if NULL, default=NULL

local_file
logical if the .nc data are available on your local disc, if FALSE the data will be downloaded, default=TRUE

file_path
string defining the path of the local file (works only if local_file = TRUE), default=NULL

sml_chunk
string defining the time period to be downloaded. Chunk available are "2011-2020", "1995-2010", "1980-1994", "1965-1979", "1950-1964"

clim_variable
string defining the daily climate variable of interest; "mean temp","max temp","min temp","precipitation", default="mean temp"

statistic
string defining the metric to retrieve, "mean" or "spread", where the mean is computed across the 100 members and is provided as the "best-guess" fields. The spread is calculated as the difference between the 5th and 95th percentiles over the ensemble to provide a measure indicate of the 90% uncertainty range. For more details see Cornes et al. (2018) and the guidance on how to use ensemble datasets available from http://surfobs.climate.copernicus.eu/userguidance/use_ensembles.php

grid_size
numeric value in degree defining the resolution of the grid, 0.25 (ca. 27 kilometres in latitude) or 0.1 (ca. 11 kilometres in latitude), default=0.25

Details

This function ask you to select the .nc file containing the data of interest from your local disc, if local_file is FALSE, data will be downloaded from the ECAD. If first_year and last_year are not provided, the function extract the full data set. Smaller chunks of about 15 years of the most recent version of the E-OBS dataset can be specified for download can be specified directly with the argument "sm_chunk" (period available are 2011-2020, 1995-2010, 1980-1994, 1965-1979, 1950-1964).

Author

Reto Schmucki